Fix indexing error
The length of `past_renewal_rate_vec` is 1 too long, causing run_simple_sim as is break. The reason for this is that fil_stats_df contains data from start_date to current_date. However, current_date is the start of simulation.  So, we need to remove one element from the vector. This fixes that.

Signed-off-by: Kiran Karra <kiran.karra@gmail.com>